backend/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Cors
pytest
pytest-flask
Flask-Migrate
sqlalchemy-libsql
pytest
pytest-flask
sphinx
sphinx_rtd_theme
flask-security
apscheduler
flask-smorest